program file
One of the challenging and unpredictable activities that has a big influence on human society is rainfall prediction. Proactively reducing human and financial loss may be achieved by timely and accurate forecasting. This article provides a series of experiments that employ popular machine learning techniques to build models that can forecast the likelihood of rain or not in major Australian cities based on the meteorological data for that particular day.